Organize files to follow logical hierchy and seperate concerns
