+++ src/homePage/stylesheets/desktop/arrows.scss
