+++ src/homePage/stylesheets/desktop/slides.scss
