+++ src/homePage/stylesheets/mobile-tablets/footersectionMobile.css
@@ -1,117 +1,3 @@
-.site-footer {
-  background-color: #131313;
-  padding: 45px 0 20px;
-  font-size: 15px;
-  line-height: 24px;
-  color: #737373;
-  margin-top: 80px;
-}
-
-.container {
-  width: 80%;
-  margin: auto;
-}
-
-.row {
-  display: flex;
-  gap: 30px;
-  justify-content: space-between;
-  place-content: space-between;
-}
-
-.column-one-text {
-  flex: 40%;
-  text-align: left;
-}
-
-.column-one-link {
-  flex: 20%;
-  text-align: left;
-}
-
-.column-two {
-  flex: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  text-align: left;
-}
-
-.social-icons {
-  width: 100%;
-  justify-content: flex-end;
-}
-
-.site-footer hr {
-  border-top: #bbb;
-  border-color: rgb(197, 168, 1);
-  opacity: 0.5;
-}
-
-.site-footer hr.small {
-  margin: 20px 0;
-}
-
-.site-footer h6 {
-  color: #fff;
-  font-size: 16px;
-  text-transform: uppercase;
-  margin-top: 5px;
-  letter-spacing: 2px;
-}
-
-.site-footer a {
-  color: #737373;
-  text-decoration: none;
-}
-
-.site-footer a:hover {
-  color: rgb(197, 168, 1);
-  text-decoration: none;
-}
-
-.footer-links {
-  padding-left: 0;
-  list-style: none;
-}
-
-.footer-links li {
-  display: block;
-}
-
-.footer-links a {
-  color: #737373;
-}
-
-.footer-links a:active,
-.footer-links a:focus,
-.footer-links a:hover {
-  color: rgb(197, 168, 1);
-  text-decoration: none;
-}
-
-.footer-links.inline li {
-  display: inline-block;
-}
-
-.site-footer .social-icons {
-  text-align: right;
-}
-
-.site-footer .social-icons a {
-  width: 40px;
-  height: 40px;
-  line-height: 40px;
-  margin-left: 6px;
-  margin-right: 0;
-  border-radius: 100%;
-  background-color: #33353d;
-}
-
-.copyright-text {
-  margin: 0;
-}
-
 @media (max-width: 991px) {
   .site-footer [class^="col-"] {
     margin-bottom: 30px;
