+++ src/homePage/stylesheets/mobile-tablets/infosectionMobile.css
