+++ src/homePage/stylesheets/mobile-tablets/slidesMobile.scss
